feature: Connecting Patents, Projects, and OtherWorks routes. colav/impactu#282 colav/impactu#281 colav/impactu#280
Create SearchPatentsPage, SearchProjectsPage and SearchOtherWorks components
Those components are responsible for displaying a list of items for this new entities based on the provided search parameters. The component uses the Suspense component to lazy load the Lists component and displays a loading indicator while the list is being fetched. The new SearchPages component improves the search functionality and user experience for patents, projects and otherWorks searches.